Renamed files to include entity
