+++ src/modules/study-definition/entities/component-definition.entity.ts
