+++ src/modules/study-definition/entities/enrollment-condition.entity.ts
