+++ src/modules/study-definition/entities/participation-criteria.entity.ts
